Refactor gamemaker app into modules
Extract pool-order comparators, swap buckets, export payload builder,
grid placement factory, and shift host from app.js. Add Node tests for
pool ordering and export payload. Keep gamemaker.html entry on app.js.
